Migración a Maven completa, incluida estructura de directorios. Se han integrado las dependencias de MySQL-Connector-9.2.0, Hibernate-core-6.6.13.Final, JUnit-5.12.2
