+++ src/main/java/modelo/enums/TipoCliente.java
@@ -1,0 +1,8 @@
+// Grupo 2 - SQL SQUAD
+
+
+package modelo.enums;
+
+public enum TipoCliente {
+    PREMIUM, ESTANDAR;
+}
