+++ src/main/java/modelo/enums/TipoEstado.java
@@ -1,0 +1,8 @@
+// Grupo 2 - SQL SQUAD
+
+
+package modelo.enums;
+
+public enum TipoEstado {
+    ENVIADO, PENDIENTE;
+}
