--- src/modelo/enums/TipoEstado.java
@@ -1,8 +1,0 @@
-// Grupo 2 - SQL SQUAD
-
-
-package modelo.enums;
-
-public enum TipoEstado {
-    ENVIADO, PENDIENTE;
-}
